apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-allauth>=65.11.2",
    "django-cors-headers>=4.8.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.10",
]
